Finish conversion to new UI format
