+++ src/ATGUI/Tabs/aimbottab.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "../atgui.h"
+
+namespace Aimbot
+{
+	void RenderTab();
+}
