+++ src/ATGUI/Tabs/hvhtab.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "../atgui.h"
+
+namespace HvH
+{
+	void RenderTab();
+}
