+++ src/ATGUI/Tabs/misctab.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "../atgui.h"
+
+namespace Misc
+{
+	void RenderTab();
+}
