+++ src/ATGUI/Tabs/modelstab.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "../atgui.h"
+
+namespace Models
+{
+	void RenderTab();
+}
