+++ src/ATGUI/Tabs/skinstab.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "../atgui.h"
+
+namespace Skins
+{
+	void RenderTab();
+}
