+++ src/ATGUI/Tabs/triggerbottab.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "../atgui.h"
+
+namespace Triggerbot
+{
+	void RenderTab();
+}
