+++ src/ATGUI/Tabs/visualstab.h
@@ -1,0 +1,8 @@
+#pragma once
+
+#include "../atgui.h"
+
+namespace Visuals
+{
+	void RenderTab();
+}
