+++ src/ATGUI/atgui.h
@@ -12,6 +12,13 @@
 #include "Windows/playerlist.h"
 #include "Windows/colors.h"
 #include "Windows/main.h"
+#include "Tabs/aimbottab.h"
+#include "Tabs/triggerbottab.h"
+#include "Tabs/visualstab.h"
+#include "Tabs/hvhtab.h"
+#include "Tabs/misctab.h"
+#include "Tabs/modelstab.h"
+#include "Tabs/skinstab.h"
 
 namespace UI
 {
